Add custom validator rule for slugs
there is the need for checking if a given string from request is
a slug or not, meaning it doeesn't contain any spaecs.
in order to have this validation function in the validator rules,
I had to create a new Rule Object called Slug, which makes a slug
version of the value, and then checks if the slug version is equal to
the value itself.