Refactor controllers folder to be more organized
there was controllers for post, category, and auth related controllers.
it is better to move each of them to a folder with the name of the resource,
so the PostController moved to post folder.
the CategoryController moved to category folder.
all related controllers related to auth moved to auth folder.